[lib] support E2EE DMs only for Vodozemac clients
Summary:
Solution #3 from [ENG-11545](https://linear.app/comm/issue/ENG-11545/handle-olm-x3dh-bug).

I had to move `useGetDeviceListsForUsers` because of dependency cycle.

Test Plan: Set `NEXT_CODE_VERSION` for something different and test what `useUsersSupportThickThreads` returns for devices during searching users

Reviewers: ashoat
