+++ lib/hooks/peer-list-hooks.js
@@ -3,6 +3,7 @@
 import invariant from 'invariant';
 import * as React from 'react';
 
+import { useGetDeviceListsForUsers } from './use-get-device-lists-for-users.js';
 import { setPeerDeviceListsActionType } from '../actions/aux-user-actions.js';
 import { usePeerOlmSessionsCreatorContext } from '../components/peer-olm-session-creator-provider.react.js';
 import {
@@ -14,23 +15,21 @@ import { usePeerToPeerCommunication } from '../tunnelbroker/peer-to-peer-context
 import { useTunnelbroker } from '../tunnelbroker/tunnelbroker-context.js';
 import { useResendPeerToPeerMessages } from '../tunnelbroker/use-resend-peer-to-peer-messages.js';
 import type {
-  UsersRawDeviceLists,
-  UsersDevicesPlatformDetails,
-  SignedDeviceList,
   RawDeviceList,
+  SignedDeviceList,
   UserDevicesPlatformDetails,
+  UsersRawDeviceLists,
 } from '../types/identity-service-types.js';
 import {
   type DeviceListUpdated,
   peerToPeerMessageTypes,
 } from '../types/tunnelbroker/peer-to-peer-message-types.js';
 import {
-  userActionsP2PMessageTypes,
   type AccountDeletionP2PMessage,
+  userActionsP2PMessageTypes,
 } from '../types/tunnelbroker/user-actions-peer-to-peer-message-types.js';
 import { getConfig } from '../utils/config.js';
 import { getContentSigningKey } from '../utils/crypto-utils.js';
-import { convertSignedDeviceListsToRawDeviceLists } from '../utils/device-list-utils.js';
 import { identityServiceQueryTimeout } from '../utils/identity-service.js';
 import { values } from '../utils/objects.js';
 import { useDispatch, useSelector } from '../utils/redux-utils.js';
@@ -41,30 +40,6 @@ export type PrimaryDeviceChange = {
   +prevPrimaryDeviceID: string,
   +newPrimaryDeviceID: string,
 };
-
-function useGetDeviceListsForUsers(): (
-  userIDs: $ReadOnlyArray<string>,
-) => Promise<{
-  +deviceLists: UsersRawDeviceLists,
-  +usersPlatformDetails: UsersDevicesPlatformDetails,
-}> {
-  const client = React.useContext(IdentityClientContext);
-  const identityClient = client?.identityClient;
-  invariant(identityClient, 'Identity client should be set');
-  return React.useCallback(
-    async (userIDs: $ReadOnlyArray<string>) => {
-      const peersDeviceLists =
-        await identityClient.getDeviceListsForUsers(userIDs);
-      return {
-        deviceLists: convertSignedDeviceListsToRawDeviceLists(
-          peersDeviceLists.usersSignedDeviceLists,
-        ),
-        usersPlatformDetails: peersDeviceLists.usersDevicesPlatformDetails,
-      };
-    },
-    [identityClient],
-  );
-}
 
 async function throwOnTimeout(userIDs: $ReadOnlyArray<string>) {
   await sleep(identityServiceQueryTimeout);
@@ -350,7 +325,6 @@ function useResetRatchetState(): (userID: ?string) => Promise<void> {
 }
 
 export {
-  useGetDeviceListsForUsers,
   useBroadcastDeviceListUpdates,
   useGetAndUpdateDeviceListsForUsers,
   useBroadcastAccountDeletion,
