+++ lib/hooks/user-identities-hooks.js
@@ -3,6 +3,7 @@
 import invariant from 'invariant';
 import * as React from 'react';
 
+import { useGetDeviceListsForUsers } from './use-get-device-lists-for-users.js';
 import { useFindUserIdentities } from '../actions/find-user-identities-actions.js';
 import { logTypes, useDebugLogs } from '../components/debug-logs-context.js';
 import { extractFIDFromUserID } from '../shared/id-utils.js';
@@ -10,6 +11,7 @@ import { IdentityClientContext } from '../shared/identity-client-context.js';
 import type { FarcasterUser } from '../types/identity-service-types.js';
 import type { AccountUserInfo } from '../types/user-types.js';
 import { getMessageForException } from '../utils/errors.js';
+import { olmSupportRequired } from '../utils/olm-support-required.js';
 import { useSelector } from '../utils/redux-utils.js';
 import { useIsFarcasterDCsIntegrationEnabled } from '../utils/services-utils.js';
 
@@ -17,6 +19,7 @@ function useUsersSupportThickThreads(): (
   userIDs: $ReadOnlyArray<string>,
 ) => Promise<$ReadOnlyMap<string, boolean | void>> {
   const findUserIdentities = useFindUserIdentities();
+  const getDeviceListsForUsers = useGetDeviceListsForUsers();
   const auxUserInfos = useSelector(state => state.auxUserStore.auxUserInfos);
 
   return React.useCallback(
@@ -25,20 +28,37 @@ function useUsersSupportThickThreads(): (
 
       const usersNeedingFetch = [];
       for (const userID of userIDs) {
-        if (auxUserInfos[userID]?.deviceList) {
-          usersSupportingThickThreads.set(userID, true);
+        const peerDeviceList = auxUserInfos[userID]?.deviceList?.devices;
+        const peerPlatformDetails =
+          auxUserInfos[userID]?.devicesPlatformDetails;
+        if (peerDeviceList && peerPlatformDetails) {
+          const supportsVodozemac = peerDeviceList.every(
+            deviceID => !olmSupportRequired(peerPlatformDetails?.[deviceID]),
+          );
+          usersSupportingThickThreads.set(userID, supportsVodozemac);
         } else {
           usersNeedingFetch.push(userID);
         }
       }
       if (usersNeedingFetch.length > 0) {
-        const { identities, reservedUserIdentifiers } =
-          await findUserIdentities(usersNeedingFetch);
+        const [
+          { identities, reservedUserIdentifiers },
+          { deviceLists, usersPlatformDetails },
+        ] = await Promise.all([
+          findUserIdentities(usersNeedingFetch),
+          getDeviceListsForUsers(usersNeedingFetch),
+        ]);
+
         for (const userID of usersNeedingFetch) {
           const isReserved = !!reservedUserIdentifiers[userID];
           const doesNotExist = identities[userID] === undefined && !isReserved;
           if (identities[userID]) {
-            usersSupportingThickThreads.set(userID, true);
+            const peerDeviceList = deviceLists[userID]?.devices ?? [];
+            const peerPlatformDetails = usersPlatformDetails[userID];
+            const supportsVodozemac = peerDeviceList.every(
+              deviceID => !olmSupportRequired(peerPlatformDetails?.[deviceID]),
+            );
+            usersSupportingThickThreads.set(userID, supportsVodozemac);
           } else if (doesNotExist) {
             usersSupportingThickThreads.set(userID, undefined);
           } else {
@@ -48,7 +68,7 @@ function useUsersSupportThickThreads(): (
       }
       return usersSupportingThickThreads;
     },
-    [auxUserInfos, findUserIdentities],
+    [auxUserInfos, findUserIdentities, getDeviceListsForUsers],
   );
 }
